Rename test listeners directory
Having the name be in the plural didn't make sense when all others are
in the singular. So, this change corrects that, renaming the directory
and the accompanying namespace references.
