+++ test/EventManager/Container/ListenerConfigurationDelegatorTest.php
@@ -7,8 +7,8 @@ namespace EventManagerIntegrationTest\Container;
 use EventManagerIntegration\Container\ListenerConfigurationDelegator;
 use EventManagerIntegration\Container\ServiceNotFoundException;
 use EventManagerIntegrationTest\InMemoryContainer;
-use EventManagerIntegrationTest\Listeners\FakeLoggerListener;
-use EventManagerIntegrationTest\Listeners\FakeNotificationListener;
+use EventManagerIntegrationTest\Listener\FakeLoggerListener;
+use EventManagerIntegrationTest\Listener\FakeNotificationListener;
 use Laminas\EventManager\EventManager;
 use PHPUnit\Framework\TestCase;
 use ReflectionClass;
